Move layercopy to util
