+++ engine/filesync/localfs.go
@@ -18,8 +18,8 @@ import (
 	"github.com/dagger/dagger/internal/buildkit/util/bklog"
 	"github.com/dagger/dagger/internal/fsutil"
 	"github.com/dagger/dagger/internal/fsutil/types"
-	"github.com/dagger/dagger/internal/layercopy"
 	"github.com/dagger/dagger/util/hashutil"
+	"github.com/dagger/dagger/util/layercopy"
 	digest "github.com/opencontainers/go-digest"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
