+++ util/layercopy/copy_linux.go
