+++ util/layercopy/types.go
